remote control tested
In this version we have:
-tested packet send and receive to set speeds
-we have fixed the measurement and speed setting to fit the orientation
of the motors on the car. So now when we say move forward, both move
forward

+++ Firmware/RaceCar/RaceCar/SerialComm.h
@@ -13,8 +13,8 @@
 #define BUFFER_SIZE 512
 #define PCKT_QUEUE_SIZE 8
 
-extern int speed_l_desired;
-extern int speed_r_desired;
+extern int8_t speed_l_desired;
+extern int8_t speed_r_desired;
 extern uint8_t motor_enable;
 extern uint8_t kp;
 extern uint8_t kd;
